+++ util/bloom_test.cc
@@ -21,7 +21,7 @@ int main() {
 #include "logging/logging.h"
 #include "memory/arena.h"
 #include "rocksdb/filter_policy.h"
-#include "table/full_filter_bits_builder.h"
+#include "table/block_based/filter_policy_internal.h"
 #include "test_util/testharness.h"
 #include "test_util/testutil.h"
 #include "util/gflags_compat.h"
